requirements.txt
streamlit==1.32.0
numpy==1.26.4
Pillow==10.0.0
matplotlib==3.7.1
opencv-python-headless==4.8.0.74
reportlab==4.0.4
tflite-runtime==2.14.0